Moving all pinot-2.0 classes to org.apache.pinot.stream.kafka20
