+++ pinot-plugins/pinot-stream-ingestion/pinot-kafka-2.0/src/main/java/org/apache/pinot/plugin/stream/kafka20/KafkaPartitionLevelConsumer.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.core.realtime.impl.kafka2;
+package org.apache.pinot.plugin.stream.kafka20;
 
 import com.google.common.collect.Iterables;
 import java.io.IOException;
