+++ pinot-plugins/pinot-stream-ingestion/pinot-kafka-2.0/src/main/resources/META-INF/services/org.apache.pinot.spi.stream.StreamConsumerFactory
@@ -16,4 +16,4 @@
 # specific language governing permissions and limitations
 # under the License.
 #
-org.apache.pinot.core.realtime.impl.kafka2.KafkaConsumerFactory
+org.apache.pinot.plugin.stream.kafka20.KafkaConsumerFactory
