+++ pinot-plugins/pinot-stream-ingestion/pinot-kafka-2.0/src/test/java/org/apache/pinot/plugin/stream/kafka20/utils/MiniKafkaCluster.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.core.realtime.impl.kafka2.utils;
+package org.apache.pinot.plugin.stream.kafka20.utils;
 
 import java.io.Closeable;
 import java.io.IOException;
